added terrain generation tool + mesh vertex data packing

+++ src/graphics/mesh/mesh.h
@@ -5,7 +5,7 @@
 
 struct MeshData
 {
-    std::vector<float> vertices;
+    std::vector<uint32_t> vertices;
     std::vector<unsigned int> indices;
 };
